requirements.txt
# LiveInsight+ Requirements
# Complete Kafka-based streaming system with ML, XAI, and Agentic AI
# Python 3.9+ recommended

# Core Kafka & Streaming
kafka-python==2.0.2
confluent-kafka==2.3.0

# Web Frameworks
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.1

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Machine Learning
scikit-learn==1.3.2
joblib==1.3.2

# Explainable AI (XAI)
shap==0.43.0
lime==0.2.0.1

# Visualization
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.0

# System Monitoring
psutil==5.9.6

# HTTP Requests
requests==2.31.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Development & Testing (optional)
pytest==7.4.3
black==23.11.0
flake8==6.1.0

# Windows-specific (optional for system monitoring)
wmi==1.5.1; sys_platform == 'win32'
pywin32==306; sys_platform == 'win32'
